requirements.txt
Flask
sqlalchemy
psycopg2-binary
pandas
numpy
pyyaml